Admin ModCrew, Booking process, Modify Booking...

+++ backend/CloudNineBackend/src/main/java/com/group12/CloudNineBackend/boundary/CrewRepo.java
@@ -7,7 +7,7 @@ import com.group12.CloudNineBackend.domain.Crew;
  * @author Jeremy Sugimoto
  *
  */
-public interface CrewRepo extends JpaRepository<Crew, Integer> {
+public interface CrewRepo extends JpaRepository<Crew, Long> {
 	void deleteById(Long id);
 
 }
